+++ src/3rdparty/walkontable/src/table/bottom.js
@@ -1,5 +1,6 @@
 import OverlayTable from './_base';
 import stickyRowsBottom from './mixin/stickyRowsBottom';
+import calculatedColumns from './mixin/calculatedColumns';
 import { mixin } from './../../../../helpers/object';
 
 class BottomOverlayTable extends OverlayTable {
@@ -7,5 +8,6 @@ class BottomOverlayTable extends OverlayTable {
 }
 
 mixin(BottomOverlayTable, stickyRowsBottom);
+mixin(BottomOverlayTable, calculatedColumns);
 
 export default BottomOverlayTable;
