+++ src/3rdparty/walkontable/src/table/left.js
@@ -1,38 +1,13 @@
 import OverlayTable from './_base';
 import calculatedRows from './mixin/calculatedRows';
+import stickyColumnsLeft from './mixin/stickyColumnsLeft';
 import { mixin } from './../../../../helpers/object';
 
 class LeftOverlayTable extends OverlayTable {
-  getFirstRenderedColumn() {
-    const totalColumns = this.wot.getSetting('totalColumns');
-    if (totalColumns === 0) {
-      return -1;
-    }
-    return 0;
-  }
 
-  getFirstVisibleColumn() {
-    return this.getFirstRenderedColumn();
-  }
-
-  getLastRenderedColumn() {
-    return this.getRenderedColumnsCount() - 1;
-  }
-
-  getLastVisibleColumn() {
-    return this.getLastRenderedColumn();
-  }
-
-  getRenderedColumnsCount() {
-    const totalColumns = this.wot.getSetting('totalColumns');
-    return Math.min(this.wot.getSetting('fixedColumnsLeft'), totalColumns);
-  }
-
-  getVisibleColumnsCount() {
-    return this.getRenderedColumnsCount();
-  }
 }
 
 mixin(LeftOverlayTable, calculatedRows);
+mixin(LeftOverlayTable, stickyColumnsLeft);
 
 export default LeftOverlayTable;
