+++ src/3rdparty/walkontable/src/table/master.js
@@ -1,5 +1,6 @@
 import Table from '../table';
 import calculatedRows from './mixin/calculatedRows';
+import calculatedColumns from './mixin/calculatedColumns';
 import { mixin } from './../../../../helpers/object';
 
 class MasterTable extends Table {
@@ -7,5 +8,6 @@ class MasterTable extends Table {
 }
 
 mixin(MasterTable, calculatedRows);
+mixin(MasterTable, calculatedColumns);
 
 export default MasterTable;
