+++ src/3rdparty/walkontable/src/table/top.js
@@ -1,5 +1,6 @@
 import OverlayTable from './_base';
 import stickyRowsTop from './mixin/stickyRowsTop';
+import calculatedColumns from './mixin/calculatedColumns';
 import { mixin } from './../../../../helpers/object';
 
 class TopOverlayTable extends OverlayTable {
@@ -7,5 +8,6 @@ class TopOverlayTable extends OverlayTable {
 }
 
 mixin(TopOverlayTable, stickyRowsTop);
+mixin(TopOverlayTable, calculatedColumns);
 
 export default TopOverlayTable;
